Refactor builtin aggregation strategies into modules
