+++ projects/04-llm-adapter/adapter/core/aggregation/builtin/max_score.py
@@ -1,0 +1,44 @@
+"""最高スコア選択ストラテジ。"""
+from __future__ import annotations
+
+from collections.abc import Sequence
+
+from .. import AggregationCandidate, AggregationResult, TieBreaker
+from .tie_breakers import FirstTieBreaker
+
+__all__ = ["MaxScoreStrategy"]
+
+
+class MaxScoreStrategy:
+    name = "max_score"
+
+    def aggregate(
+        self, candidates: Sequence[AggregationCandidate], *, tiebreaker: TieBreaker | None = None
+    ) -> AggregationResult:
+        if not candidates:
+            raise ValueError("max_score: candidates must be non-empty")
+
+        if any(candidate.score is not None for candidate in candidates):
+            chosen = max(
+                candidates,
+                key=lambda c: (c.score is not None, float(c.score or float("-inf")), -c.index),
+            )
+            return AggregationResult(
+                chosen=chosen,
+                candidates=list(candidates),
+                strategy=self.name,
+                reason=f"score={chosen.score}",
+                tie_breaker_used=None,
+                metadata=None,
+            )
+
+        breaker = tiebreaker or FirstTieBreaker()
+        chosen = breaker.break_tie(candidates)
+        return AggregationResult(
+            chosen=chosen,
+            candidates=list(candidates),
+            strategy=self.name,
+            reason="all scores are None → tie-break",
+            tie_breaker_used=breaker.name,
+            metadata=None,
+        )
