+++ projects/04-llm-adapter/adapter/core/aggregation/builtin/tie_breakers.py
@@ -1,0 +1,31 @@
+"""組み込みタイブレーカー。"""
+from __future__ import annotations
+
+from collections.abc import Sequence
+
+from .. import AggregationCandidate
+
+__all__ = ["FirstTieBreaker", "MaxScoreTieBreaker"]
+
+
+class FirstTieBreaker:
+    name = "stable_order"
+
+    def break_tie(self, candidates: Sequence[AggregationCandidate]) -> AggregationCandidate:
+        if not candidates:
+            raise ValueError("TieBreaker: candidates must be non-empty")
+        return min(candidates, key=lambda c: c.index)
+
+
+class MaxScoreTieBreaker:
+    name = "max_score"
+
+    def break_tie(self, candidates: Sequence[AggregationCandidate]) -> AggregationCandidate:
+        if not candidates:
+            raise ValueError("TieBreaker: candidates must be non-empty")
+        if any(c.score is not None for c in candidates):
+            return max(
+                candidates,
+                key=lambda c: (c.score is not None, float(c.score or float("-inf")), -c.index),
+            )
+        return FirstTieBreaker().break_tie(candidates)
